[#314] 'CircuitBreaker.isFailure' logic not being executed.
The current logic which checks 'isFailure' in 'CircuitBreakerImpl.apply'

'case Left(e) if isFailure.isDefinedAt(e) => ZIO.fail(e)'

isn't accomplishing the desired result and in almost all cases will return 'false'.  The reason is that 'isDefinedAt' only checks if 'e' is covered within the domain of the 'PartialFunction', however it does not execute the logic of the partial function.

instead 'applyOrElse' should be used to perform both operations:

'case Left(e) if isFailure.applyOrElse[E1, Boolean](e, _ => false) => ZIO.fail(e)'

There is also a bug in the 'CircuitBreakerSpec' that is preventing the test from failing (the 'either' is in the wrong spot).

Housekeeping:
- ZIO updated to 2.0.8